issue50: introduced DTOs for data from frontend
