--- de.dashup.shared/src/main/java/de/dashup/shared/DashupPanelInformation.java
@@ -1,23 +1,0 @@
-package de.dashup.shared;
-
-class DashupPanelInformation {
-
-    private String id;
-    private int panel_order;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public int getPanelOrder() {
-        return panel_order;
-    }
-
-    public void setPanelOrder(int panel_order) {
-        this.panel_order = panel_order;
-    }
-}
