+++ de.dashup.shared/src/main/java/de/dashup/shared/DataTransferObject.java
@@ -1,0 +1,4 @@
+package de.dashup.shared;
+
+public interface DataTransferObject {
+}
